requirements.txt
asgiref 
astroid==2.5
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django
django-cors-headers==3.8.0
django-filter==2.4.0
django-import-export==2.5.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
et-xmlfile==1.1.0
idna==3.2
isort==5.7.0
jwcrypto==1.0
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
psycopg2
pycparser==2.20
PyJWT==2.1.0
pylint==2.7.1
python-decouple==3.4
python3-openid==3.2.0
pytz
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rope==0.18.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse
tablib==3.0.0
toml==0.10.2
typing-extensions
urllib3==1.26.6
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
